file: use 64bit indexing for block size
This patch changes the block size and block compression functions to use
64bit indexing. From reading the specs, I came to the conclusion that
the block size table can be larger than 4Gi entries. (to be exact, I
believe that the worst case block table is:

	UINT64_MAX (max file size) / 4096 (minimal block size)

While we don't support files that large yet, we lay the foundation to
support them in the future.

+++ libsqsh/include/sqsh_file_private.h
@@ -147,7 +147,7 @@ struct SqshFileIterator {
 	struct SqshFragmentView fragment_view;
 	size_t sparse_size;
 	size_t block_size;
-	uint32_t block_index;
+	uint64_t block_index;
 	const uint8_t *data;
 	size_t size;
 };
